+++ CHANGES.md
@@ -1,3 +1,10 @@
 # Changes
+
+## Version 2020013000
+
+- Fixed bug stripping id tags from templates.
+- Added a setting to let Moodle clean the template HTML or output the raw template with no checks.
+
 ## Version 2017103000
-*   Added support for Font-Awesome icon mapping.
+
+- Added support for Font-Awesome icon mapping.